# services/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Xora Waypoint API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/stores/me/orders": {
      "get": {
        "tags": [
          "store"
        ],
        "summary": "Home",
        "operationId": "home_api_v1_stores_me_orders_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreHomeOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/orders/{ref}": {
      "get": {
        "tags": [
          "store"
        ],
        "summary": "Get Order",
        "operationId": "get_order_api_v1_orders__ref__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ref"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/outlets/{code}/usual-items": {
      "get": {
        "tags": [
          "store"
        ],
        "summary": "Usual Items",
        "operationId": "usual_items_api_v1_outlets__code__usual_items_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UsualItemOut"
                  },
                  "title": "Response Usual Items Api V1 Outlets  Code  Usual Items Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/orders/check": {
      "post": {
        "tags": [
          "store"
        ],
        "summary": "Check",
        "operationId": "check_api_v1_orders_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCheckIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderCheckOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/orders": {
      "post": {
        "tags": [
          "store"
        ],
        "summary": "Submit",
        "operationId": "submit_api_v1_orders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderSubmitIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrderOut"
                  },
                  "type": "array",
                  "title": "Response Submit Api V1 Orders Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/outlets/{code}/driver-note": {
      "post": {
        "tags": [
          "store"
        ],
        "summary": "Driver Note",
        "operationId": "driver_note_api_v1_outlets__code__driver_note_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverNoteIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverNoteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications": {
      "get": {
        "tags": [
          "store"
        ],
        "summary": "Notifications",
        "operationId": "notifications_api_v1_notifications_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NotificationOut"
                  },
                  "type": "array",
                  "title": "Response Notifications Api V1 Notifications Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/notifications/{notice_id}": {
      "get": {
        "tags": [
          "store"
        ],
        "summary": "Notification",
        "operationId": "notification_api_v1_notifications__notice_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "notice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Notice Id"
            }
          },
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "en",
                    "si",
                    "ta"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Language"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/{notice_id}/ack": {
      "post": {
        "tags": [
          "store"
        ],
        "summary": "Acknowledge",
        "operationId": "acknowledge_api_v1_notifications__notice_id__ack_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "notice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Notice Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/{notice_id}/read": {
      "post": {
        "tags": [
          "store"
        ],
        "summary": "Acknowledge",
        "operationId": "acknowledge_api_v1_notifications__notice_id__read_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "notice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Notice Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{ref}/receipt-draft": {
      "get": {
        "tags": [
          "receipts"
        ],
        "summary": "Draft",
        "operationId": "draft_api_v1_orders__ref__receipt_draft_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ref"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptDraftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{ref}/receipt": {
      "post": {
        "tags": [
          "receipts"
        ],
        "summary": "Confirm",
        "operationId": "confirm_api_v1_orders__ref__receipt_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ref"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceiptIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{ref}/issues": {
      "post": {
        "tags": [
          "receipts"
        ],
        "summary": "Report",
        "operationId": "report_api_v1_orders__ref__issues_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ref"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{ref}/issues/check": {
      "post": {
        "tags": [
          "receipts"
        ],
        "summary": "Check Report",
        "operationId": "check_report_api_v1_orders__ref__issues_check_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ref"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuePreviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DriverNoteIn": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 240,
            "minLength": 1,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "DriverNoteIn"
      },
      "DriverNoteOut": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "sent_at": {
            "type": "string",
            "format": "date-time",
            "title": "Sent At"
          }
        },
        "type": "object",
        "required": [
          "text",
          "sent_at"
        ],
        "title": "DriverNoteOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NotificationOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "order_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Ref"
          },
          "reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code"
          },
          "lang": {
            "type": "string",
            "title": "Lang"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "next_run": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Run"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "acked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acked At"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "Written from the plan"
          },
          "protection": {
            "type": "string",
            "title": "Protection",
            "default": "First in line (moved once)"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "order_ref",
          "reason_code",
          "lang",
          "title",
          "body",
          "next_run",
          "created_at",
          "acked_at"
        ],
        "title": "NotificationOut"
      },
      "OrderCheckIn": {
        "properties": {
          "lines": {
            "items": {
              "$ref": "#/components/schemas/ItemIn"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Lines"
          }
        },
        "type": "object",
        "required": [
          "lines"
        ],
        "title": "OrderCheckIn"
      },
      "OrderCheckOut": {
        "properties": {
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/QuantityWarning"
            },
            "type": "array",
            "title": "Warnings"
          },
          "delivery_date": {
            "type": "string",
            "format": "date",
            "title": "Delivery Date"
          },
          "cutoff_at": {
            "type": "string",
            "format": "date-time",
            "title": "Cutoff At"
          },
          "cutoff_seconds": {
            "type": "integer",
            "title": "Cutoff Seconds"
          },
          "after_cutoff": {
            "type": "boolean",
            "title": "After Cutoff"
          }
        },
        "type": "object",
        "required": [
          "warnings",
          "delivery_date",
          "cutoff_at",
          "cutoff_seconds",
          "after_cutoff"
        ],
        "title": "OrderCheckOut"
      },
...
      "OrderOut": {
        "properties": {
          "ref": {
            "type": "string",
            "title": "Ref"
          },
          "outlet_code": {
            "type": "string",
            "title": "Outlet Code"
          },
          "brand": {
            "type": "string",
            "title": "Brand"
          },
          "district": {
            "type": "string",
            "title": "District"
          },
          "temp_requirement": {
            "type": "string",
            "title": "Temp Requirement"
          },
          "delivery_date": {
            "type": "string",
            "format": "date",
            "title": "Delivery Date"
          },
          "units": {
            "type": "integer",
            "title": "Units"
          },
          "status": {
            "type": "string",
            "enum": [
              "planned",
              "loaded",
              "in_transit",
              "delivered",
              "at_risk",
              "late",
              "deferred",
              "pending_sync",
              "on_hold"
            ],
            "title": "Status"
          },
          "submission_status": {
            "type": "string",
            "title": "Submission Status"
          },
          "placed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Placed At"
          },
          "receipt_confirmed": {
            "type": "boolean",
            "title": "Receipt Confirmed"
          },
          "receipt_confirmed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Receipt Confirmed At"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/OrderLineOut"
            },
            "type": "array",
            "title": "Lines"
          },
          "tracking": {
            "items": {
              "$ref": "#/components/schemas/TrackingOut"
            },
            "type": "array",
            "title": "Tracking"
          }
        },
        "type": "object",
        "required": [
          "ref",
          "outlet_code",
          "brand",
          "district",
          "temp_requirement",
          "delivery_date",
          "units",
          "status",
          "submission_status",
          "placed_at",
          "receipt_confirmed",
          "receipt_confirmed_at",
          "lines",
          "tracking"
        ],
        "title": "OrderOut"
      },
      "OrderSubmitIn": {
        "properties": {
          "lines": {
            "items": {
              "$ref": "#/components/schemas/ItemIn"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Lines"
          },
          "draft_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft Ref"
          },
          "confirm_unusual": {
            "type": "boolean",
            "title": "Confirm Unusual",
            "default": false
          },
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "lines",
          "request_id"
        ],
        "title": "OrderSubmitIn"
      },
...
      "ReceiptIn": {
        "properties": {
          "lines": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ReceiptCountIn"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lines"
          }
        },
        "type": "object",
        "title": "ReceiptIn"
      },
      "ReceiptOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "order_ref": {
            "type": "string",
            "title": "Order Ref"
          },
          "total_cases": {
            "type": "integer",
            "title": "Total Cases"
          },
          "confirmed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Confirmed At"
          },
          "issue_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issue Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "order_ref",
          "total_cases",
          "confirmed_at"
        ],
        "title": "ReceiptOut"
      },
...
      "StoreHomeOut": {
        "properties": {
          "outlet_code": {
            "type": "string",
            "title": "Outlet Code"
          },
          "district": {
            "type": "string",
            "title": "District"
          },
          "brand": {
            "type": "string",
            "title": "Brand"
          },
          "now": {
            "type": "string",
            "format": "date-time",
            "title": "Now"
          },
          "ordering": {
            "$ref": "#/components/schemas/OrderCheckOut"
          },
          "window_open": {
            "type": "string",
            "format": "time",
            "title": "Window Open"
          },
          "window_close": {
            "type": "string",
            "format": "time",
            "title": "Window Close"
          },
          "orders": {
            "items": {
              "$ref": "#/components/schemas/OrderOut"
            },
            "type": "array",
            "title": "Orders"
          },
          "driver_note": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DriverNoteOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "outlet_code",
          "district",
          "brand",
          "now",
          "ordering",
          "window_open",
          "window_close",
          "orders",
          "driver_note"
        ],
        "title": "StoreHomeOut"
      },
...
      "UsualItemOut": {
        "properties": {
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "is_chilled": {
            "type": "boolean",
            "title": "Is Chilled"
          },
          "usual_qty": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usual Qty"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "name",
          "is_chilled",
          "usual_qty"
        ],
        "title": "UsualItemOut"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}